backend/requirements.txt
fastapi
uvicorn
sqlalchemy
python-multipart
python-jose
passlib[bcrypt]
pydantic